Simple physic feature

+++ game/ecs/phy/Velocity.js
@@ -1,6 +1,6 @@
 import { defineComponent, Types } from '../../bitecs.js';
 
 export const Velocity = defineComponent({
-    x: Types.i16,
-    y: Types.i16
+    x: Types.f32,
+    y: Types.f32
 });
